requirements.txt
# Core
asyncio-mqtt>=2.0
aiohttp>=3.9
websockets>=12.0
pydantic>=2.5
pyyaml>=6.0

# Polymarket / Blockchain
py-clob-client>=0.15
web3>=6.15
eth-account>=0.11

# Data & Analysis
numpy>=1.26
pandas>=2.2
scipy>=1.12

# AI / LLM
anthropic>=0.40

# News & Data Feeds
feedparser>=6.0
tweepy>=4.14
aiohttp-sse-client>=0.2

# Monitoring & Logging
structlog>=24.1
prometheus-client>=0.20
rich>=13.7

# Storage
aiosqlite>=0.19
redis>=5.0
